Move SessionType to base module
