+++ freqtrade/persistence/base.py
@@ -1,6 +1,8 @@
 
-from sqlalchemy.orm import DeclarativeBase
+from sqlalchemy.orm import DeclarativeBase, Session, scoped_session
 
+
+SessionType = scoped_session[Session]
 
 class ModelBase(DeclarativeBase):
     pass
